requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
discord.py==1.2.3
humanize==0.5.1
idna==2.8
import-expression==0.3.8.post2
jishaku==1.15.0.134
multidict==4.5.2
numpy==1.17.0
psutil==5.6.3
websockets==6.0
yarl==1.3.0
